+++ docs/components/setup/sections/SectionSetup.vue
@@ -1,6 +1,7 @@
 <template>
 <section id='setup' class='section'>
   <div class='container'>
+    <h3 class='title has-text-primary is-spaced'>Setup</h3>
     <b-message type='is-warning'>
       <a href='https://vuejs.org' target='_blank'>Vue.js</a> version <strong>2.4.2+</strong> is required.
     </b-message>