+++ src/components/CalendarPane.vue
@@ -1,6 +1,6 @@
 <template>
   <div
-    class='c-pane'>
+    :class='["c-pane", { "is-full-width": position === 0 }]'>
     <!--Header-->
     <div class='c-header-wrapper'>
       <!--Header vertical divider-->
@@ -23,43 +23,45 @@
             </slot>
           </div>
           <!--Header title-->
-          <popover
-            align='center'
-            :class='["c-title", titleClass]'
-            :style='{ width: "100%" }'>
-            <transition-group
-              tag='div'
-              :name='titleTransition_'>
-              <div
-                class='c-title-1'
-                v-for='p in pages'
-                :key='p.key'
-                v-if='p === page_'>
-                <div class='c-title-2'>
-                  <slot name='header-title' :page='p'>
-                    <div class='c-title-span' :style='titleStyle'>
-                      <span @click='navMode = "month"'>
-                        {{ p.monthLabel }}
-                      </span>
-                      <span @click='navMode = "year"'>
-                        {{ p.yearLabel }}
-                      </span>
-                    </div>
+          <div
+            :class='["c-title-layout", titleClass]'>   
+            <!--Navigation popover--> 
+            <popover
+              class='c-title-popover'
+              visibility='hover'
+              :align='titlePosition'
+              :content-style='{ padding: "0" }'
+              :enter-delay='300'>
+              <!--Title content-->
+              <transition-group
+                tag='div'
+                class='c-title-anchor'
+                :name='titleTransition_'>
+                <div
+                  class='c-title'
+                  :style='titleStyle'
+                  v-for='p in pages'
+                  :key='p.key'
+                  v-if='p === page_'
+                  @click='$emit("titleClick", p)'>
+                  <slot
+                    name='header-title'>
+                    {{ `${p.monthLabel} ${p.yearLabel}` }}
                   </slot>
                 </div>
-              </div>
-            </transition-group>
-            <!--Navigation pane-->
-            <calendar-nav
-              slot='popover-content'
-              :mode='navMode'
-              :monthLabels='monthLabels'
-              :value='page_'
-              :min-page='minPage'
-              :max-page='maxPage'
-              @input='move($event)'>
-            </calendar-nav>
-          </popover>
+              </transition-group>
+              <!--Navigation pane-->
+              <calendar-nav
+                slot='popover-content'
+                :mode.sync='navMode'
+                :monthLabels='monthLabels'
+                :value='page_'
+                :validator='canMove'
+                :attributes='attributes'
+                @input='move($event)'>
+              </calendar-nav>
+            </popover>
+          </div>
           <!--Header next button-->
           <div class='c-arrow-layout'>
             <slot name='header-right-button' :page='page_'>
@@ -133,10 +135,11 @@
             :next-month-comps='p.nextMonthComps'
             :first-day-of-week='firstDayOfWeek'
             :styles='styles'
+            :attributes='attributes'
             v-bind='$attrs'
-            @touchstart='touchStart($event)'
-            @touchmove='touchMove($event)'
-            @touchend='touchEnd($event)'
+            @touchstart.passive='touchStart($event)'
+            @touchmove.passive='touchMove($event)'
+            @touchend.passive='touchEnd($event)'
             v-on='$listeners'
             v-if='p === page_'>
           </calendar-weeks>
@@ -165,6 +168,7 @@ import {
 import {
   monthLabels,
   weekdayLabels,
+  titlePosition,
   titleTransition,
   weeksTransition,
   maxSwipeTimeMs,
@@ -187,9 +191,10 @@ export default {
     weekdayLabels: { type: Array, default: () => weekdayLabels },
     firstDayOfWeek: { type: Number, default: 1 },
     styles: Object,
-    titlePosition: String,
+    titlePosition: { type: String, default: titlePosition },
     titleTransition: { type: String, default: titleTransition },
     weeksTransition: { type: String, default: weeksTransition },
+    attributes: Array,
   },
   data() {
     return {
@@ -371,7 +376,7 @@ export default {
       // Preload other pages
       this.preloadPages();
     },
-    loadPage({ month, year }, position = 0) {
+    loadPage({ month, year }) {
       const key = `${year.toString()}.${month.toString()}`;
       let page = this.pages.find(p => (p.key === key));
       if (!page) {
@@ -408,15 +413,15 @@ export default {
         };
         this.pages.push(page);
       }
-      page.position = position;
+      page.position = this.position;
       page.loaded = true;
       return page;
     },
     preloadPages() {
       // Load the next and previous pages
       Vue.nextTick(() => {
-        this.loadPage(this.page_.prevMonthComps, -1);
-        this.loadPage(this.page_.nextMonthComps, 1);
+        this.loadPage(this.page_.prevMonthComps);
+        this.loadPage(this.page_.nextMonthComps);
         this.pages = this.pages.filter(p => p.loaded);
         this.pages.forEach((p) => {
           p.loaded = false;
@@ -452,14 +457,13 @@ export default {
 @import '../styles/mixins.sass'
 
 .c-pane
-  flex-grow: 1
-  flex-shrink: 1
-  // min-width: $paneMinWidth
-  // width: $paneWidth
+  width: 50%
   display: flex
   flex-direction: column
   align-items: stretch
   overflow: hidden
+  &.is-full-width
+    width: 100%
 
 .c-header-wrapper
   display: flex
@@ -483,41 +487,30 @@ export default {
       user-select: none
       &:hover
         opacity: 0.5
-    
-  .c-title
-    +box()
+  .c-title-layout
+    display: inline-flex
+    justify-content: center
+    align-items: center
     flex-grow: 1
-    position: relative
-    .c-title-1
-      position: absolute
-      left: 0
-      top: 0
-      width: 100%
-      height: 100%
+    .c-title-popover
       display: flex
-      align-items: center
-      .c-title-2
-        +box()
-        width: 100%
-        .c-title-span
+      justify-content: inherit
+      .c-title-anchor
+        display: flex
+        justify-content: inherit
+        .c-title
           font-weight: $titleFontWeight
           font-size: $titleFontSize
+          transition: $titleTransition
+          cursor: pointer
           user-select: none
-          text-align: center
-          span
-            transition: $titleTransition
-            cursor: pointer
-            &:hover
-              opacity: 0.5
-
+          white-space: nowrap
     &.align-left
       order: -1
-      .c-title-2
-        justify-content: flex-start
+      justify-content: flex-start
     &.align-right
       order: 1
-      .c-title-2
-        justify-content: flex-end
+      justify-content: flex-end
     
   .c-arrow.c-disabled
     cursor: not-allowed
@@ -558,17 +551,18 @@ export default {
   flex-direction: column
   width: 100%
 
-.title-slide-left-enter-active,
 .title-slide-left-leave-active,
-.title-slide-right-enter-active,
 .title-slide-right-leave-active,
-.title-slide-up-enter-active,
 .title-slide-up-leave-active,
-.title-slide-down-enter-active,
 .title-slide-down-leave-active,
-.title-fade-enter-active,
 .title-fade-leave-active
-  transition: $titleTransition
+  position: absolute
+
+.title-slide-left-leave,
+.title-slide-right-leave,
+.title-slide-up-leave,
+.title-slide-down-leave
+  opacity: 1
 
 .title-none-enter-active,
 .title-none-leave-active