+++ src/components/DatePicker.vue
@@ -54,9 +54,9 @@ import Popover from './Popover';
 import SingleDatePicker from './SingleDatePicker';
 import MultipleDatePicker from './MultipleDatePicker';
 import DateRangePicker from './DateRangePicker';
-import { DateInfo, blendColors } from '../utils/helpers';
+import DateInfo from '../utils/dateInfo';
+import { blendColors } from '../utils/helpers';
 
-const POPOVER_AUTO = -1;
 const _defaultSelectColor = '#66b3cc';
 const _defaultDragColor = '#9fcfdf';
 
@@ -74,7 +74,7 @@ export default {
     isExpanded: Boolean,
     popoverDirection: { type: String, default: 'bottom' },
     popoverAlign: { type: String, default: 'left' },
-    popoverVisibility: { type: Number, default: POPOVER_AUTO },
+    popoverVisibility: { type: String, default: 'hover' },
     inputClass: String,
     inputStyle: Object,
     inputPlaceholder: String,